--- src/main/java/fly/be/flyflix/auth/controller/TokenController.java
@@ -1,23 +1,0 @@
-package fly.be.flyflix.auth.controller;
-
-import fly.be.flyflix.auth.controller.dto.LoginRequest;
-import fly.be.flyflix.auth.controller.dto.LoginResponse;
-import fly.be.flyflix.auth.service.LoginService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TokenController {
-
-    @Autowired
-    private LoginService loginService;
-
-
-    @PostMapping("/login")
-    public ResponseEntity <LoginResponse> login(@RequestBody LoginRequest loginRequest) {
-        return loginService.login(loginRequest);
-    }
-}
